+++ packages/iov-bns/src/index.ts
@@ -39,6 +39,7 @@ export {
   VoteOption,
   CreateProposalTx,
   isCreateProposalTx,
+  Vote,
   VoteTx,
   isVoteTx,
   // Proposals