+++ packages/iov-bns/types/index.d.ts
@@ -36,6 +36,7 @@ export {
   VoteOption,
   CreateProposalTx,
   isCreateProposalTx,
+  Vote,
   VoteTx,
   isVoteTx,
   ProposalAction,