+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.16.2
 
+- @iov/bns: Add `getVotes` by voter address method to `BnsConnection` class.
+- @iov/bns-governance: Add `getVotes` method to `Governor` class.
 - @iov/crypto: Add `EnglishMnemonic.wordlist`.
 
 ## 0.16.1